fix: dev server config
